Rename airfilter.py to airfilter_util.py to indicate it's not referring to an actual device
