+++ miio/airpurifier.py
@@ -5,7 +5,7 @@ from typing import Any, Dict, Optional
 
 import click
 
-from .airfilter import FilterType, FilterTypeUtil
+from .airfilter_util import FilterType, FilterTypeUtil
 from .click_common import EnumType, command, format_output
 from .device import Device
 from .exceptions import DeviceException
